chore(page): Update result page
