+++ src/pages/questions/paper/index.jsx
@@ -1,17 +1,63 @@
 import Taro, { Component } from '@tarojs/taro';
 import { View } from '@tarojs/components';
 import StepPage from "@/components/StepPage"
+import { connect } from "@tarojs/redux";
 import { AtTextarea } from 'taro-ui'
 import './index.scss';
 
+@connect(({ user, step, loading }) => ({
+    user,
+    step,
+    loading
+}))
 export default class Paper extends Component {
     state = {
         value: "",
         placeholder: "经过以上实验，同学你有什么想法呢？清在这里撰写你的实验报告，字数 150 字以内。"
     }
 
-    handleNextClick = (callback) => {
+    handleNextClick = async (callback) => {
+        if (this.state.value == null || this.state.value == "") {
+            Taro.showToast({
+                icon: 'none',
+                title: "请填写实验报告"
+            })
+            callback(false)
+            return
+        }
+        await this.endExperiment()
+        await this.saveExperimentLog()
         callback(true)
+    }
+
+    async endExperiment() {
+        await this.props
+            .dispatch({
+                type: "step/endExperiment",
+                payload: {
+                    userExperiment: Object.assign(this.props.step.userExperiment, {
+                        log: JSON.stringify(this.props.step.stepQueue.exportAll()),
+                        paper: this.state.value
+                    })
+                }
+            })
+            .then(() => { });
+    }
+
+    saveExperimentLog() {
+        this.props
+            .dispatch({
+                type: "step/saveExperimentLog",
+                payload: {
+                    params: this.props.step.userExperiment
+                }
+            })
+            .then(() => {
+                Taro.showToast({
+                    icon: "none",
+                    title: "提交试验记录成功"
+                });
+            });
     }
 
     handleChange(value) {
@@ -22,8 +68,13 @@ export default class Paper extends Component {
 
     render() {
         const { value, placeholder } = this.state
+        const isSaveExperimentLogLoading = this.props.loading.effects[
+            "step/saveExperimentLog"
+        ];
         return (
-            <StepPage onNext={this.handleNextClick} showPanel>
+            <StepPage onNext={this.handleNextClick} showPanel
+                nextButtonLoading={isSaveExperimentLogLoading}
+            >
                 <View className='paper-container'>
                     <AtTextarea
                         value={value}
